Moved components to a separate file to make testing easier
