+++ client/src/ErrorMessageDiv.js
@@ -1,5 +1,7 @@
 import React, { Component }  from 'react';
 
+// Stateless Higher Order Component for displaying the No user found message at the top of the screen
+//
 //https://codeburst.io/displaying-error-messages-in-react-with-a-higher-order-component-hoc-fe2de074bf64
 const WithErrorHandling = WrappedComponent => ({ showError, children }) => {
     return (
